estel/requirements.txt
langchain==0.3.23
langchain-cohere==0.4.4
langchain-community==0.3.21
langchain-core==0.3.51
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
faiss-cpu==1.10.0
streamlit==1.34.0
python-dotenv==1.0.1
tqdm==4.66.2
pypdf==3.17.4
cryptography>=3.1
